move useBlockchainEventFilter to a separate file
