+++ frontend/src/components/wallet/useWallet.ts
@@ -1,7 +1,5 @@
 import { WalletContext } from 'components/wallet/WalletContext';
-import { useEffect, useContext } from 'react';
-import { BaseContract, EventFilter } from 'ethers';
-import { Listener } from '@ethersproject/providers';
+import { useContext } from 'react';
 
 export const useWallet = () => {
   const context = useContext(WalletContext);
@@ -9,19 +7,4 @@ export const useWallet = () => {
     throw new Error('No Contract Context Found');
   }
   return context;
-};
-
-export const useBlockchainEventFilter = (
-  contract: BaseContract,
-  filter: EventFilter,
-  handlerCallback: Listener
-) => {
-  useEffect(() => {
-    // attach
-    contract.on(filter, handlerCallback);
-    return () => {
-      // detach
-      contract.off(filter, handlerCallback);
-    };
-  }, [contract, filter, handlerCallback]);
 };
